artxctic-backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# AWS S3 / Cloudflare R2
boto3==1.36.4
botocore==1.36.4

# Redis
redis==5.2.1

# Background Tasks
celery==5.4.0
APScheduler==3.10.4

# Google OAuth
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# Dodopayments
dodopayments[webhooks]>=1.81.0

# Logging & Monitoring
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2

# Utils
email-validator==2.2.0
